+++ Frends.RabbitMQ.Read/Frends.RabbitMQ.Read/Read.cs
@@ -1,7 +1,10 @@
 ﻿using System;
 using System.Collections.Generic;
 using System.ComponentModel;
+using System.Diagnostics.CodeAnalysis;
 using System.Linq;
+using System.Security.Authentication;
+using System.Security.Cryptography.X509Certificates;
 using System.Text;
 using System.Threading.Tasks;
 using Frends.RabbitMQ.Read.Definitions;
@@ -26,7 +29,7 @@ public class RabbitMQ
         var baseList = new List<Message>();
         var stringList = new List<Message>();
 
-        await OpenConnectionIfClosed(connectionHelper, connection);
+        await Helper.OpenConnectionIfClosed(connectionHelper, connection);
 
         while (connection.ReadMessageCount-- > 0)
         {
@@ -36,7 +39,7 @@ public class RabbitMQ
                 baseList.Add(new Message
                 {
                     Data = Convert.ToBase64String(rcvMessage.Body.ToArray()),
-                    Headers = GetResponseHeaderDictionary(rcvMessage.BasicProperties),
+                    Headers = Helper.GetResponseHeaderDictionary(rcvMessage.BasicProperties),
                     MessagesCount = rcvMessage.MessageCount,
                     DeliveryTag = rcvMessage.DeliveryTag
                 });
@@ -44,7 +47,7 @@ public class RabbitMQ
                 stringList.Add(new Message
                 {
                     Data = Encoding.UTF8.GetString(rcvMessage.Body.ToArray()),
-                    Headers = GetResponseHeaderDictionary(rcvMessage.BasicProperties),
+                    Headers = Helper.GetResponseHeaderDictionary(rcvMessage.BasicProperties),
                     MessagesCount = rcvMessage.MessageCount,
                     DeliveryTag = rcvMessage.DeliveryTag
                 });
@@ -92,90 +95,6 @@ public class RabbitMQ
             case AckType.AutoAck:
             default:
                 throw new ArgumentException($"AcknowledgeMessage should not be called with {ackType}.", nameof(ackType));
-        }
-    }
-
-    private static Dictionary<string, string> GetResponseHeaderDictionary(IReadOnlyBasicProperties basicProperties)
-    {
-        if (basicProperties == null) return null;
-
-        var allHeaders = new Dictionary<string, string>()
-            {
-                { "HEADER_APPID",             basicProperties.AppId != null ? basicProperties.AppId : null },
-                { "HEADER_CLUSTERID",         basicProperties.ClusterId != null ? basicProperties.ClusterId : null },
-                { "HEADER_CONTENTENCODING",   basicProperties.ContentEncoding != null ? basicProperties.ContentEncoding : null },
-                { "HEADER_CONTENTTYPE",       basicProperties.ContentType != null ? basicProperties.ContentType : null },
-                { "HEADER_CORRELATIONID",     basicProperties.CorrelationId != null ? basicProperties.CorrelationId : null },
-                { "HEADER_EXPIRATION",        basicProperties.Expiration != null ? basicProperties.Expiration : null},
-                { "HEADER_MESSAGEID",         basicProperties.MessageId != null ? basicProperties.MessageId : null }
-            }
-        .Where(h => h.Value != null)
-        .ToDictionary(h => h.Key, h => h.Value);
-
-        if (basicProperties.IsHeadersPresent())
-            foreach (var header in basicProperties.Headers.ToList())
-            {
-                if (header.Value.GetType() == typeof(byte[]))
-                    allHeaders[header.Key] = Encoding.UTF8.GetString(header.Value as byte[]);
-                else
-                    allHeaders[header.Key] = header.Value.ToString();
-            }
-
-        return allHeaders;
-    }
-
-    private static async Task OpenConnectionIfClosed(ConnectionHelper connectionHelper, Connection connection)
-    {
-        // Close connection if hostname has changed.
-        if (IsConnectionHostNameChanged(connectionHelper, connection))
-            await connectionHelper.AMQPModel.CloseAsync();
-
-        if (connectionHelper.AMQPConnection == null || connectionHelper.AMQPConnection.IsOpen == false)
-        {
-            var factory = new ConnectionFactory();
-
-            switch (connection.AuthenticationMethod)
-            {
-                case AuthenticationMethod.URI:
-                    factory.Uri = new Uri(connection.Host);
-                    break;
-                case AuthenticationMethod.Host:
-                    if (!string.IsNullOrWhiteSpace(connection.Username) || !string.IsNullOrWhiteSpace(connection.Password))
-                    {
-                        factory.UserName = connection.Username;
-                        factory.Password = connection.Password;
-                    }
-                    factory.HostName = connection.Host;
-
-                    if (connection.Port != 0) factory.Port = connection.Port;
-
-                    break;
-            }
-
-            if (connection.Timeout != 0) factory.RequestedConnectionTimeout = TimeSpan.FromSeconds(connection.Timeout);
-
-            connectionHelper.AMQPConnection = await factory.CreateConnectionAsync();
-        }
-
-        if (connectionHelper.AMQPModel == null || connectionHelper.AMQPModel.IsClosed)
-            connectionHelper.AMQPModel = await connectionHelper.AMQPConnection.CreateChannelAsync();
-    }
-
-    private static bool IsConnectionHostNameChanged(ConnectionHelper connectionHelper, Connection connection)
-    {
-        // If no current connection, host name is not changed
-        if (connectionHelper.AMQPConnection == null || connectionHelper.AMQPConnection.IsOpen == false)
-            return false;
-
-        switch (connection.AuthenticationMethod)
-        {
-            case AuthenticationMethod.URI:
-                var newUri = new Uri(connection.Host);
-                return (connectionHelper.AMQPConnection.Endpoint.HostName != newUri.Host);
-            case AuthenticationMethod.Host:
-                return (connectionHelper.AMQPConnection.Endpoint.HostName != connection.Host);
-            default:
-                throw new ArgumentException($"IsConnectionHostNameChanged: AuthenticationMethod missing.");
         }
     }
 }